[add] simple enums

+++ include/macros.h
@@ -83,11 +83,10 @@ only the enum NAMES gets STRINGIZE'd
 #define SOTA_ZOOMTOPOINT(scale, x, origin) \
     (scale * (x - origin) + origin)
 
-/* -- Computing sign from value -- */
-// 1. (val == 0) -> Reduce any value to [0, 1].
-// 2. If val is true, returns    1
-// 3. If val is false, returns  -1
-#define SIGN(val) (-2 * (val == 0) + 1)
+/* -- Computing sign from switch -- */
+// 1. If val is true, returns    1
+// 2. If val is false, returns  -1
+#define SIGN(val) (-2 * (int)(val == 0) + 1)
 
 /* -- Error check routine --
 **  1. assert for debug
